More guidelines for task format
The LLM would occasionally give badly formatted tasks and they were not handled well.  (read: loops plan/join) and eventually the LLM gives a good task and it continues on. This commit updates prompts in several places to help give more guidance for the LLM on how to format task plans it generates.