Add comments and change a path for args to fail
Adds more comments on what's happening.

Also changes one logic path for task argument resolution to fail if it reaches that branch. Before it just passed the args back if they didn't parse at this branch, but it would be safer to fail to avoid silent or hard to trace bugs.