Fix string or list in math ctx param
The math function parameter for context only took a list, but the llm passes strings while other LangChain Runnables will pass lists. Since it's just turning lists into strings anyway, this change adds the option for the parameter to be passed as a string in addition to a list.

Other changes:
- Fixes the math example in the planner prompt
- Updates the math prompts to be more readable by making them a multiline string and tweaking some small things in it.
- Add some more examples to run

The math function still has some issues, noting in the README.